Code review for integrity check
